+++ CHANGELOG.markdown
@@ -2,6 +2,9 @@
 ### Added
 - Added event options to `track`, `track_anonymous`, and `pageview` for setting top-level event `id`, `timestamp`, and `type` fields separately from event attributes.
 
+### Deprecated
+- Passing `timestamp` in event attributes to set the top-level event timestamp is deprecated. Pass `timestamp` as an event option instead.
+
 ## Customerio 5.4.0 - June 13, 2025
 ### Changed
 - Added `send_sms` to `APIClient` and `SendSMSRequest` to support sending transactional push notifications.